Refactored msdp.hpp into msdp/variable.hpp.
That was all that was left after factoring out the protocol part.
