+++ include/telnetpp/options/msdp/server.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "telnetpp/options/msdp.hpp"
+#include "telnetpp/options/msdp/variable.hpp"
 #include "telnetpp/server_option.hpp"
 
 namespace telnetpp { namespace options { namespace msdp {
