+++ src/options/msdp/variable.cpp
@@ -1,4 +1,4 @@
-#include "telnetpp/options/msdp.hpp"
+#include "telnetpp/options/msdp/variable.hpp"
 
 namespace telnetpp { namespace options { namespace msdp {
